+++ fabric/src/main/java/dev/tapetum/shaders/mixin/MixinTerrainVertexBuffer.java
@@ -11,18 +11,20 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(VertexBuffer.class)
 public abstract class MixinTerrainVertexBuffer {
-    @Shadow private int vertextBufferId;
+    @Shadow private int vertexBufferId;
     @Shadow private int indexCount;
     @Shadow private VertexFormat.IndexType indexType;
+    @Shadow private com.mojang.blaze3d.systems.RenderSystem.AutoStorageIndexBuffer sequentialIndices;
     @Shadow private VertexFormat.Mode mode;
     @Shadow private VertexFormat format;
 
-    @Inject(method = "drawChunkLayer", at = @At(value = "INVOKE",
+    @Inject(method = "draw", at = @At(value = "INVOKE",
         target = "Lcom/mojang/blaze3d/systems/RenderSystem;drawElements(III)V"), cancellable = true)
     private void tapetum$drawTerrain(CallbackInfo ci) {
         int indices = org.lwjgl.opengl.GL11.glGetInteger(org.lwjgl.opengl.GL15.GL_ELEMENT_ARRAY_BUFFER_BINDING);
-        if (TapetumShaders.getPipelineManager().drawTerrain(vertextBufferId, indices, indexCount,
-                indexType.asGLType, format, mode)) {
+        var actualType = sequentialIndices == null ? indexType : sequentialIndices.type();
+        if (TapetumShaders.getPipelineManager().drawTerrain(vertexBufferId, indices, indexCount,
+                actualType.asGLType, format, mode)) {
             ci.cancel();
         }
     }